Securing the views product and profile views

+++ products/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.db.models.functions import Lower
 
 from .models import Product, Category
@@ -63,8 +64,13 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, this page is not authorized for you!')
+        return redirect(reverse('home'))
  
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -85,8 +91,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in webshop """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, this page is not authorized for you!')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
@@ -110,8 +121,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from webshop"""
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, this page is not authorized for you!')
+        return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
